ADD: Moved all group target blocks to a seperate file
